Introduce message queues for recieving messages and use them in the server
Synchronous message queues now store the messages, which can now be
used to recieve multiple messages at once.
We moved a lot of the application logic from onData to the server's
main() function.
How we process the message on the server side is still a bit janky, and
will need improvements.